// nest api/src/users/interfaces/user.ts
export interface UserInterface {
    id?: number;
    uid: string;
    username?: string;
    displayName?: string;
    fullname?: string;
    cedula?: string;
    email: string;
    address?: string;
    phone?: string;
    phone2?: string;
    facebook?: string;
    password: string;
    photoUrl?: string;
    createdAt?: string;
}